Rename main to content to prevent confusion.
