+++ gen-my-pod/src/components/Content.tsx
